--share-cert support

+++ run.sh
@@ -9,11 +9,12 @@ usage() {
 	echo "    -k|--hook <hook-url>      [Optional] URL to be hit before server execution, for DDNS update or notification"
 	echo "    --wp <websocket-path>     [Optional] Enable websocket with websocket-path setting, e.g. '/wsocket'. default disable"
 	echo "    --nginx <domain-name>     [Optional] Enable Ngnix frontend with given domain-name, must be applied with --wp enabled"
-	echo "    --nginx-port <port-num>   [Optional] Ngnix listening port, default 8443, must be applied with --nginx enabled"
+	echo "    --nginx-port <port-num>   [Optional] Ngnix listening port, default 443, must be applied with --nginx enabled"
+	echo "    --share-cert <cert-path>  [Optional] Waiting for cert populating in given path instead of requesting. default disable"
 	echo "    --no-ssl                  [Optional] Disable Ngnix SSL support for CDN optimisation, must be applied with --nginx enabled"
 }
 
-TEMP=`getopt -o u:p:l:a:k: --long uuid:,port:,level:,alterid:,hook:,wp:,nginx:,nginx-port:,nginx-ssl -n "$0" -- $@`
+TEMP=`getopt -o u:p:l:a:k: --long uuid:,port:,level:,alterid:,hook:,wp:,nginx:,nginx-port:,share-cert:,nginx-ssl -n "$0" -- $@`
 if [ $? != 0 ] ; then usage; exit 1 ; fi
 
 eval set -- "$TEMP"
@@ -40,8 +41,13 @@ while true ; do
 			shift 2
 			;;
 		--wp)
-			WSPATH="$2"
-			shift 2
+			if [[ $2 =~ ^\/[A-Za-z0-9_-]{1,16}$ ]]; then
+				WSPATH="$2"
+				shift 2
+			else
+				echo "Websocket path must be 1-16 aplhabets, numbers, '-' or '_' started with '/'"
+				exit
+			fi
 			;;
 		--nginx)
 			NGDOMAIN="$2"
@@ -49,6 +55,10 @@ while true ; do
 			;;
 		--nginx-port)
 			NGPORT="$2"
+			shift 2
+			;;
+		--share-cert)
+			SHARECERT="$2"
 			shift 2
 			;;
 		--no-ssl)
@@ -84,7 +94,7 @@ if [ -z "${LEVEL}" ]; then
 fi
 
 if [ -z "${NGPORT}" ]; then
-	NGPORT=8443
+	NGPORT=443
 fi
 
 if [ -n "${HOOKURL}" ]; then
@@ -92,41 +102,62 @@ if [ -n "${HOOKURL}" ]; then
 	echo
 fi
 
+if [ -z "${SHARECERT}" ]; then
+	CERTPATH="/root/.acme.sh/${NGDOMAIN}"
+else
+	CERTPATH="${SHARECERT}"
+fi
+
 if [ -n "${NGDOMAIN}" ]; then
 	if [ -z "${WSPATH}" ]; then
-		echo "'--wp' option is missing, which is necessary for nginx frontend running. Abrot."
+		echo "'--wp' option is missing, which is necessary for '--nginx' option. Abrot."
 		exit 1
 	fi
 
 	if [ -z "${NOSSL}" ]; then
 		TRY=0
-		while [ ! -f "/root/.acme.sh/${NGDOMAIN}/fullchain.cer" ]
+		while [ ! -f "${CERTPATH}/fullchain.cer" ]
 		do
-			/root/.acme.sh/acme.sh --issue --standalone -d ${NGDOMAIN}
-			((TRY++))
-			if [ $TRY >= 3 ]; then
-				echo "Obtian cert for ${NGDOMAIN} failed. Check log please."
-				exit 3
+			if [ -n "${SHARECERT}" ]; then
+				echo "Cert populating not found, Waitting..."
+			else
+				echo "Cert requesting..."
+				/root/.acme.sh/acme.sh --issue --standalone -d ${NGDOMAIN}
+				((TRY++))
+				if [ ${TRY} >= 3 ]; then
+					echo "Requesting cert for ${NGDOMAIN} failed. Check log please."
+					exit 3
+				fi
 			fi
+			echo "Wait 10 seconds before cert checking again..."
+			sleep 10
 		done
 	fi
 
+	# Running as root to enable low port listening. Necessary for Fargate or k8s.
 	sed -i 's/^user nginx;$/user root;/g' /etc/nginx/nginx.conf
 	mkdir -p /run/nginx/
 
 	cd /etc/nginx/conf.d/
-	mv default.conf default.conf.disable
+
+	if [ -f /etc/nginx/conf.d/default.conf ]; then
+		mv default.conf default.conf.disable
+	fi
+
 	if [ -z "${NOSSL}" ]; then
 		TPL="site-ssl.conf.tpl"
 	else
 		TPL="site-non-ssl.conf.tpl"
 	fi
+
 	cat ${TPL} \
-		| sed 's/NGDOMAIN/${NGDOMAIN}/g' \
-		| sed 's/NGPORT/${NGPORT}/g' \
-		| sed 's/VPORT/${VPORT}/g' \
-		| sed 's/WSPATH/${WSPATH}/g' \
+		| sed "s/CERTPATH/${CERTPATH}/g" \
+		| sed "s/NGDOMAIN/${NGDOMAIN}/g" \
+		| sed "s/NGPORT/${NGPORT}/g" \
+		| sed "s/VPORT/${VPORT}/g" \
+		| sed "s/WSPATH/\\${WSPATH}/g" \
 		>v2site.conf
+
 fi
 
 cd /tmp
@@ -140,8 +171,8 @@ cat /usr/bin/v2ray/vpoint_vmess_freedom.json \
 if [ -n "${WSPATH}" ]; then
 	cat server.json \
 		| jq "(.inbounds[] | select( .protocol == \"vmess\")) +=  {\"streamSettings\":{\"network\":\"ws\",\"wsSettings\":{\"path\":\"${WSPATH}\"}}}" - \
-		>ws.json
-	mv ws.json server.json
+		>server-ws.json
+	mv server-ws.json server.json
 fi
 
 nginx
